breaking: merge @cypress/angular-zoneless upstream into @cypress/angular and remove angular support for angular 18, 19 and 20.
BREAKING CHANGE: merge @cypress/angular-zoneless upstream into @cypress/angular and remove angular support for angular 18, 19, and 20. Additionally, @cypress/schematic only supports Cypress 16 and doesnt scaffold the angular-zoneless mount handler any longer. This is a breaking change to Cypress, @cypress/angular, and @cypress/schematic. Additionally, @cypress/angular-zoneless will be deprecated on NPM due to the upstream move
i
