+++ .circleci/src/pipeline/workflows/@main.yml
@@ -241,9 +241,6 @@ linux-x64:
     - npm-angular:
         requires:
           - build
-    - npm-angular-zoneless:
-        requires:
-          - build
     - npm-mount-utils:
         requires:
           - build
@@ -265,7 +262,6 @@ linux-x64:
           - check-ts
           - health-check
           - npm-angular
-          - npm-angular-zoneless
           - npm-eslint-plugin-dev
           - npm-puppeteer-unit-tests
           - npm-puppeteer-cypress-tests
