+++ cli/.gitignore
@@ -19,7 +19,6 @@ vue
 react*
 mount-utils
 angular
-angular-zoneless
 svelte
 index.js
 lib/**/*.js
