+++ cli/eslint.config.ts
@@ -36,7 +36,6 @@ export default [
       '**/build/**/*',
       'package.json',
       '**/angular/**/*',
-      '**/angular-zoneless/**/*',
       '**/react/**/*',
       '**/vue/**/*',
       '**/svelte/**/*',
