+++ cli/package.json
@@ -70,7 +70,6 @@
   },
   "devDependencies": {
     "@cypress/angular": "0.0.0-development",
-    "@cypress/angular-zoneless": "0.0.0-development",
     "@cypress/grep": "0.0.0-development",
     "@cypress/mount-utils": "0.0.0-development",
     "@cypress/react": "0.0.0-development",
@@ -117,7 +116,6 @@
     "vue",
     "react",
     "angular",
-    "angular-zoneless",
     "svelte"
   ],
   "bin": {
@@ -158,11 +156,6 @@
       "types": "./angular/dist/index.d.ts",
       "import": "./angular/dist/index.js",
       "require": "./angular/dist/index.js"
-    },
-    "./angular-zoneless": {
-      "types": "./angular-zoneless/dist/index.d.ts",
-      "import": "./angular-zoneless/dist/index.js",
-      "require": "./angular-zoneless/dist/index.js"
     },
     "./svelte": {
       "types": "./svelte/dist/index.d.ts",
