+++ cli/scripts/bundle-ct-frameworks.ts
@@ -11,7 +11,6 @@ const npmModulesToCopy: string[] = [
   'react',
   'vue',
   'angular',
-  'angular-zoneless',
   'svelte',
 ]
