--- npm/angular-zoneless/CHANGELOG.md
@@ -1,6 +1,0 @@
-# @cypress/angular-zoneless-v1.0.0 (2025-12-05)
-
-
-### Features
-
-* add `cypress/angular-zoneless` testing harness for Angular 21 (also supports Angular 20) ([#33025](https://github.com/cypress-io/cypress/issues/33025)) ([0f215f3](https://github.com/cypress-io/cypress/commit/0f215f3f554e3f99624b25e19c03ecfd771aba52))
