--- npm/angular-zoneless/CLAUDE.md
@@ -1,3 +1,0 @@
-# @cypress/angular-zoneless
-
-@AGENTS.md
