--- npm/angular-zoneless/README.md
@@ -1,15 +1,0 @@
-# @cypress/angular-zoneless
-
-Mount Angular components in the open source [Cypress.io](https://www.cypress.io/) test runner without leveraging zone.js for change detection
-
-> **Note:** This package is bundled with the `cypress` package and should not need to be installed separately. See the [Angular Component Testing Docs](https://docs.cypress.io/guides/component-testing/angular/overview) for mounting Angular components. Installing and importing `mount` from `@cypress/angular-zoneless` should only be done for advanced use-cases.
-
-## Requirements
-
-- Angular 21.0.0 or Angular 20.0.0 with development preview of zoneless change detection
-
-## Development
-
-Run `yarn build` to compile and sync packages to the `cypress` cli package.
-
-## [Changelog](./CHANGELOG.md)
