--- npm/angular-zoneless/package.json
@@ -1,74 +1,0 @@
-{
-  "name": "@cypress/angular-zoneless",
-  "version": "0.0.0-development",
-  "description": "Test Angular Components with Cypress without zone.js",
-  "main": "dist/index.js",
-  "scripts": {
-    "prebuild": "rimraf dist",
-    "build": "rollup -c rollup.config.mjs",
-    "postbuild": "node ../../scripts/sync-exported-npm-with-cli.js",
-    "check-ts": "tsc --noEmit",
-    "dev": "rollup -c rollup.config.mjs -w",
-    "lint": "eslint --ext .js,.ts,.json, ."
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@angular/common": "^21.0.0",
-    "@angular/core": "^21.0.0",
-    "@angular/platform-browser-dynamic": "^21.0.0",
-    "@cypress/mount-utils": "0.0.0-development",
-    "rollup": "^4.60.2",
-    "typescript": "~5.9.2"
-  },
-  "peerDependencies": {
-    "@angular/common": ">=20.0.0",
-    "@angular/core": ">=20.0.0",
-    "@angular/platform-browser-dynamic": ">=20.0.0",
-    "rxjs": ">=7.8.0"
-  },
-  "files": [
-    "dist"
-  ],
-  "types": "dist/index.d.ts",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/cypress-io/cypress.git"
-  },
-  "homepage": "https://github.com/cypress-io/cypress/blob/develop/npm/angular-zoneless/#readme",
-  "bugs": "https://github.com/cypress-io/cypress/issues/new?assignees=&labels=npm%3A%20%40cypress%2Fangular-zoneless&template=1-bug-report.md&title=",
-  "keywords": [
-    "angular",
-    "cypress",
-    "cypress-io",
-    "test",
-    "testing",
-    "zoneless"
-  ],
-  "contributors": [
-    {
-      "name": "Bill Glesias",
-      "social": "@atofstryker"
-    }
-  ],
-  "module": "dist/index.js",
-  "publishConfig": {
-    "access": "public"
-  },
-  "nx": {
-    "targets": {
-      "build": {
-        "outputs": [
-          "{workspaceRoot}/cli/angular-zoneless"
-        ]
-      }
-    }
-  },
-  "standard": {
-    "globals": [
-      "Cypress",
-      "cy",
-      "expect"
-    ]
-  }
-}
