--- npm/angular-zoneless/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "experimentalDecorators": true,
-    "target": "es2022",
-    "module": "es2022",
-    "moduleResolution": "bundler",
-    "skipLibCheck": true,
-    "lib": [
-      "ESNext",
-      "DOM"
-    ],
-    "allowJs": true,
-    "declaration": true,
-    "outDir": "dist",
-    "strict": true,
-    "baseUrl": "./",
-    "types": [
-      "cypress"
-    ],
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "noPropertyAccessFromIndexSignature": true,
-  },
-  "include": ["src/**/*.*"],
-  "exclude": ["src/**/*-spec.*"]
-}
