+++ npm/angular/rollup.config.mjs
@@ -6,8 +6,6 @@ const config = {
     '@angular/core/testing',
     '@angular/common',
     '@angular/platform-browser-dynamic/testing',
-    'zone.js',
-    'zone.js/testing',
     '@angular/core/rxjs-interop',
   ],
 }
