+++ npm/cypress-schematic/src/schematics/utils/dependencies.ts
@@ -26,13 +26,6 @@ export interface NodeDependency {
   overwrite?: boolean
 }
 
-const ALL_DEPENDENCY_TYPE = [
-  NodeDependencyType.Default,
-  NodeDependencyType.Dev,
-  NodeDependencyType.Optional,
-  NodeDependencyType.Peer,
-]
-
 export function addPackageJsonDependency (
   tree: Tree,
   dependency: NodeDependency,
@@ -46,38 +39,4 @@ export function addPackageJsonDependency (
   if (overwrite || !json.get(path)) {
     json.modify(path, version)
   }
-}
-
-export function removePackageJsonDependency (
-  tree: Tree,
-  name: string,
-  pkgJsonPath = PKG_JSON_PATH,
-): void {
-  const json = new JSONFile(tree, pkgJsonPath)
-
-  for (const depType of ALL_DEPENDENCY_TYPE) {
-    json.remove([depType, name])
-  }
-}
-
-export function getPackageJsonDependency (
-  tree: Tree,
-  name: string,
-  pkgJsonPath = PKG_JSON_PATH,
-): NodeDependency | null {
-  const json = new JSONFile(tree, pkgJsonPath)
-
-  for (const depType of ALL_DEPENDENCY_TYPE) {
-    const version = json.get([depType, name])
-
-    if (typeof version === 'string') {
-      return {
-        type: depType,
-        name,
-        version,
-      }
-    }
-  }
-
-  return null
 }
