+++ npm/vue/tsconfig.json
@@ -47,7 +47,7 @@
 
     /* Experimental Options */
     // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
-    "esModuleInterop": true
+    "esModuleInterop": true,
   },
   "include": ["src", "cypress/component/**/*.{ts,tsx}", "cypress/component/**/*.vue"],
   "exclude": ["**/Translated*.vue"]
