+++ packages/scaffold-config/README.md
@@ -14,7 +14,7 @@ We will also attempt to scaffold a configuration file for projects using React a
 | React   | -          | Webpack    | 5           | React 18, 19           | `@cypress/vue@latest`     | [Link](../../system-tests/projects/react18)                        |
 | Vue     | -          | Vite       | 5, 6, 7, 8 | Vue 3                  | `@cypress/react@latest`   | [Link](../../system-tests/projects/vue3-vite-ts-configured)        |
 | Vue     | -          | Webpack    | 5           | Vue 3                  | `@cypress/vue@latest`     | [Link](../../system-tests/projects/vue3-webpack-ts-configured)     |
-| Angular | -          | Webpack    | 5           | Angular 18, 19, 20, 21 | `@cypress/angular@latest`  | [Link](../../system-tests/projects/angular-cli-configured)         |
+| Angular | -          | Webpack    | 5           | Angular 21 | `@cypress/angular@latest`  | [Link](../../system-tests/projects/angular-cli-configured)         |
 | Svelte  | -          | Vite       | 5, 6, 7, 8 | Svelte 5               | `@cypress/svelte@latest`  | [Link](../../system-tests/projects/svelte-vite-configured)         |
 | Svelte  | -          | Webpack    | 5           | Svelte 5               | `@cypress/svelte@latest`  | [Link](../../system-tests/projects/svelte-webpack-configured)      |
 | Next.js | 14, 15, 16 | Webpack    | 5           | React 18, 19           | `@cypress/react@latest`   | [Link](../../system-tests/projects/nextjs-configured)              |
