+++ packages/scaffold-config/src/frameworks.ts
@@ -101,16 +101,6 @@ export function getBundler (bundler: WizardBundler['type']): WizardBundler {
 
 const mountModule = <T extends string>(mountModule: T) => (projectPath: string) => Promise.resolve(mountModule)
 
-const angularMountModule = async (projectPath: string) => {
-  const angularCorePkg = await isDependencyInstalled(dependencies.WIZARD_DEPENDENCY_ANGULAR_CORE, projectPath)
-
-  if (!angularCorePkg.detectedVersion || !semver.valid(angularCorePkg.detectedVersion)) {
-    return 'cypress/angular'
-  }
-
-  return semver.major(angularCorePkg.detectedVersion) >= 21 ? 'cypress/angular-zoneless' : 'cypress/angular'
-}
-
 export const SUPPORT_STATUSES: Readonly<Cypress.ResolvedComponentFrameworkDefinition['supportStatus'][]> = ['alpha', 'beta', 'full', 'community'] as const
 
 export const CT_FRAMEWORKS: Cypress.ComponentFrameworkDefinition[] = [
@@ -198,7 +188,7 @@ export const CT_FRAMEWORKS: Cypress.ComponentFrameworkDefinition[] = [
     },
     codeGenFramework: 'angular',
     glob: '*.component.ts',
-    mountModule: angularMountModule,
+    mountModule: mountModule('cypress/angular'),
     supportStatus: 'full',
     componentIndexHtml: componentIndexHtmlGenerator(),
     specPattern: '**/*.cy.ts',
