+++ system-tests/projects/angular-signals/cypress/support/component.ts
@@ -1,4 +1,4 @@
-import { mount } from 'cypress/angular-zoneless'
+import { mount } from 'cypress/angular'
 
 // Augment the Cypress namespace to include type definitions for
 // your custom command.
